bug a corriger projection mauvaise + norme

+++ FdF/utils.c
@@ -6,7 +6,7 @@
 /*   By: ale-sain <ale-sain@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/19 08:15:38 by alvina            #+#    #+#             */
-/*   Updated: 2023/01/24 14:46:53 by ale-sain         ###   ########.fr       */
+/*   Updated: 2023/01/25 11:32:19 by ale-sain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ int	ft_isalnum(int c)
 		return (0);
 }
 
-void	img_pixel_put(t_data *img, double x, double y, int color)
+void	img_pixel_put(t_data *img, int x, int y, int color)
 {
 	char	*px;
